Empecé a modificar la carpeta de componentes. La idea es separarla como un repositorio independiente (SD).
