+++ src/components/atoms/A/style.module.css
@@ -1,91 +1,89 @@
 /*
   SD
   Componentes
-  Link (Enlace) _ Estilos
+  HTML a - Estilos
   José Esteva <josesteva@soft4pilot.net>
   
 */
 
 /* Predeterminado */
 
-
-
-.link {
+a {
   text-decoration: none;
 }
 
-.link img {
+a img {
   outline-style: none;
   outline-width: 3px;
 }
 
 /* Estados y eventos */
 
-.link:hover {
+a:hover {
   cursor: pointer;
 }
 
-.link span:hover {
+a span:hover {
   text-decoration: underline;
 }
 
-.link img:hover {
+a img:hover {
   outline-style: solid;
 }
 
 /* Colores */
 
-.black *:hover {
+a.black *:hover {
   color: var(--color-black);
   outline-color: var(--color-black);
 }
 
-.primary *:hover {
+a.primary *:hover {
   color: var(--color-primary);
   outline-color: var(--color-primary);
 }
 
-.primary.dark *:hover {
+a.primary.dark *:hover {
   color: var(--color-primary-dark);
   outline-color: var(--color-primary-dark);
 }
 
-.primary.light *:hover {
+a.primary.light *:hover {
   color: var(--color-primary-light);
   outline-color: var(--color-primary-light);
 }
 
-.secondary *:hover {
+a.secondary *:hover {
   color: var(--color-secondary);
   outline-color: var(--color-secondary);
 }
 
-.secondary.dark *:hover {
+a.secondary.dark *:hover {
   color: var(--color-secondary-dark);
   outline-color: var(--color-secondary-dark);
 }
 
-.secondary.light *:hover {
+a.secondary.light *:hover {
   color: var(--color-secondary-light);
   outline-color: var(--color-secondary-light);
 }
 
-.neutral *:hover {
+a.neutral *:hover {
   color: var(--color-neutral);
   outline-color: var(--color-neutral);
 }
 
-.neutral.dark *:hover {
+a.neutral.dark *:hover {
   color: var(--color-neutral-dark);
   outline-color: var(--color-neutral-dark);
 }
 
-.neutral.light *:hover {
+a.neutral.light *:hover {
   color: var(--color-neutral-light);
   outline-color: var(--color-neutral-light);
 }
 
-.white *:hover {
+a.white *:hover {
   color: var(--color-white);
   outline-color: var(--color-white);
 }
