+++ src/components/atoms/A/styleguide.mdx
@@ -2,7 +2,7 @@
   SD
   Componentes
   Guía de estilo
-  Link (Enlace) - Descripción
+  HTML `<a>` (Enlace) - Descripción
   José Esteva <josesteva@soft4pilot.net>
 */}
 
@@ -11,7 +11,7 @@ import * as Stories from './styleguide.stories';
 
 <Meta of={Stories} />
 
-# Link (Enlace)
+# HTML `<a>` (Enlace)
 
 Lorem ipsum dolor sit amet consectetur, adipisicing elit. Pariatur qui ut
 beatae nam molestiae cum doloremque fugit nihil, vitae modi recusandae culpa
