+++ src/components/molecules/Link/index.test.jsx
@@ -1,12 +1,12 @@
 // SD
 // Componentes
-// Link (Enlace) - Pruebas
+// Link (Etiqueta) - Pruebas
 // José Esteva <josesteva@soft4pilot.net>
 // 
 
 import { render } from '@testing-library/react';
 import Link from './index';
 
-test('Componente <Link/>', () => {
-  render(<Link/>);
+test('Componente <Link role="h1"/>', () => {
+  render(<Link>Etiqueta</Link>);
 });
