+++ src/styles.css
@@ -8,8 +8,8 @@
 
 @import url('https://fonts.googleapis.com/css2?family=Alegreya+Sans:wght@400;700&family=Croissant+One&display=swap');
 
-@import 'styles/tokens/style.css';
-@import 'styles/default/style.css';
+@import 'components/styles/tokens/style.css';
+@import 'components/styles/default/style.css';
 
 /* [/lock:global-styles] */
